Attempt to fix missing numbers
Occasionally an xml file will have a number missing. This ends up as a
message to a blank number. The output then has a .html because the
format is <number>.html but the number is blank
This commit should fix that under the assumption that the missing number
is the carrier number (added some logic to confirm this). If that's the
case, the missing number can be dropped.

This may also inadvertantly fix the issue of self messages not being
processed. The previous logic seemed to assume that there was always a
separate sender and receiver. If both are the carrier, the message will
be silently dropped. This is no longer the case. The message will be
stored in <carrier_number>.html.